requirements.txt
asgiref==3.5.2
attrs==21.4.0
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
Django==4.0.6
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-jsonapi==5.0.0
filter==0.0.0.20200724
gunicorn==20.1.0
html5lib==1.1
idna==3.3
inflection==0.5.1
iniconfig==1.1.1
packaging==21.3
psycopg2==2.9.3
py==1.11.0
pyparsing==3.0.7
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.1
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
tomli==2.0.1
urllib3==1.26.11
webencodings==0.5.1
